add additional fixes to make PetIBM with petsc3.5 and 3.6
makefile was looking for `variables` in a path beginning
$PETSC_DIR/$PETSC_ARCH but it doesn't have $PETSC_ARCH in the path for
either 3.5.4 or 3.6.1

adding a new variables to configure_petsc.m4 to point to the variables
file for either version

updated makefiles with changes to m4

fix ioPetIBM to work with either petsc 3.5 or 3.6

ioPetIBM needs to import PetscBinaryIO.py from $PETSC_DIR, but the path
to that file changed between versions.  There is a petsc routine to
return the version of petsc being used, but rather than use that, this
is a simple fix that tries to import the file from the 3.6 path first,
and if that throws an exception, it tries to use the 3.5 path.

resolves #65