Relax the restriction of the version of OpenMPI
Relax the restriction of the version of OpenMPI. Now OpenMPI with
version larger than 1.8 can also be used. However, users need to create
a symbolic link called "libmpi.so.1" to the mpi library they use. This
is because the AmgX released by NVIDIA was compiled with libmpi.so.1,
and we have nothing to do with that.