Correct SolvType to SolveType
In simulationParameters.yaml, we used to use "pSolvType" and "vSolvType" to
specify whether solvers will run on GPU or CPU. "SolvType" in these two
parameter is actually a typo of "SolveType". Now we correct it. So now users
have to use "pSolveType" and "vSolveType" to specify whether solvers will run
on GPU or CPU.